+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/batch/TransactionalBulkExecutor.java
@@ -830,6 +830,7 @@ public final class TransactionalBulkExecutor implements Disposable {
             docClientWrapper,
             container,
             cosmosBatchBulkOperation.getPartitionKeyValue(),
+            null,
             null);
     }